Restructured xml utils to dedicated package.
