+++ .gitignore
@@ -4,3 +4,5 @@
 /.classpath
 /.project
 /.settings
+**/.DS_Store
+
