+++ src/test/resources/org/smack/util/xml/cdd_did.xsl
